H2 database support. All update queries added.

+++ src/main/java/org/kacper/rental_items/Accessory.java
@@ -4,7 +4,7 @@ public class Accessory extends RentalItem {
     private final String description;
     private final AccessoryType type;
     
-    public Accessory(Long id, Double pricePerHour, String name, String description, AccessoryType type) {
+    public Accessory(int id, Double pricePerHour, String name, String description, AccessoryType type) {
         super(id, pricePerHour, name);
         this.description = description;
         this.type = type;
