+++ src/main/java/org/kacper/rental_items/RentalItem.java
@@ -2,17 +2,17 @@ package org.kacper.rental_items;
 
 
 public abstract class RentalItem {
-    private final Long id;
+    private final int id;
     private final Double pricePerHour;
     private final String name;
     
-    protected RentalItem(Long id, Double pricePerHour, String name) {
+    protected RentalItem(int id, Double pricePerHour, String name) {
         this.id = id;
         this.pricePerHour = pricePerHour;
         this.name = name;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
